requirements.txt
# Web Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0

# OCR Engine - Use compatible versions
paddlepaddle>=3.0.0,<3.1.0
paddleocr>=3.3.0
imgaug==0.4.0
numpy<2.0.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Validation
pydantic==2.5.0
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
